--- webview-ui/src/page/Home.tsx
@@ -1,8 +1,0 @@
-function Home() {
-  return (
-    <h1>heloow
-    </h1>
-  );
-}
-
-export default Home;
