+++ webview-ui/src/types.ts
@@ -1,0 +1,6 @@
+export interface PackageSource {
+  title: string;
+  path: string;
+  checked: boolean;
+  selected: boolean;
+}
